Renames and change module locations
