--- src/config.rs
@@ -1,2 +1,0 @@
-pub mod db_settings;
-pub mod settings;
