+++ src/configurations.rs
@@ -1,0 +1,4 @@
+pub mod settings;
+pub mod db_settings;
+pub mod application_settings;
+pub mod variables;
