+++ src/configurations/application_settings.rs
@@ -1,0 +1,11 @@
+use serde::Deserialize;
+use serde_aux::field_attributes::deserialize_number_from_string;
+
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct ApplicationSettings {
+    #[serde(deserialize_with="deserialize_number_from_string")]
+    pub port: u16,
+    pub host: String,
+    pub base_url: String,
+}
