+++ src/configurations/settings.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize};
+
+use super::db_settings::DatabaseSettings;
+
+
+
+#[derive(Debug, Deserialize, Clone)]
+pub struct Settings {
+    pub database: DatabaseSettings,
+    
+}
