+++ src/controllers/handle_redirect.rs
@@ -4,7 +4,8 @@ use redis::{Client as RedisClient};
 use crate::{helpers::{
     response::{TResult, ApiBody, make_request, ResponseBuilder}, 
     request::{HyperClient}, keyval::KeyVal, commons::GrantType}, 
-    setup::variables::SettingsVars, errors::response::{TError}, middlewares::request_builder::{RequestBuilder, AuthType}, interceptor::handle_request::{Interceptor, V2TokensType}, startup::server::AppState
+    configurations::variables::SettingsVars, errors::response::{TError}, middlewares::request_builder::{RequestBuilder, AuthType}, 
+    interceptors::handle_request::{Interceptor, V2TokensType}, startup::server::AppState
 };
