+++ src/controllers/revoke_token.rs
@@ -3,7 +3,8 @@ use serde::{Serialize, Deserialize};
 
 use crate::{helpers::{
     keyval::KeyVal, response::{TResult, ApiBody, make_request, ResponseBuilder}}, 
-    setup::variables::SettingsVars, middlewares::request_builder::{RequestBuilder, AuthType}, interceptor::handle_request::Interceptor, startup::server::AppState
+    configurations::variables::SettingsVars, middlewares::request_builder::{RequestBuilder, AuthType}, 
+    interceptors::handle_request::Interceptor, startup::server::AppState
 };
