+++ src/lib.rs
@@ -5,5 +5,5 @@ pub mod helpers;
 pub mod setup;
 pub mod routes;
 pub mod middlewares;
-pub mod interceptor;
-pub mod config;
+pub mod interceptors;
+pub mod configurations;
