--- src/setup.rs
@@ -1,3 +1,0 @@
-pub mod variables;
-// pub mod credentials;
-// pub mod client;
